+++ README.md
@@ -4,10 +4,15 @@
 
 2 단계 요구사항
 
-- [ ] HTTP Reponse의 상태 응답 코드를 302로 반환한다.
-- [ ] POST로 들어온 요청의 Request Body를 파싱할 수 있다.
-- [ ] 로그인에 성공하면 HTTP Reponse의 헤더에 Set-Cookie가 존재한다.
-- [ ] 서버에 세션을 관리하는 클래스가 있고, 쿠키로부터 전달 받은 JSESSIONID 값이 저장된다.
+- [x] HTTP Reponse의 상태 응답 코드를 302로 반환한다.
+- [x] POST로 들어온 요청의 Request Body를 파싱할 수 있다.
+- [x] 로그인에 성공하면 HTTP Reponse의 헤더에 Set-Cookie가 존재한다.
+- [x] 서버에 세션을 관리하는 클래스가 있고, 쿠키로부터 전달 받은 JSESSIONID 값이 저장된다.
+
+3 단계 요구사항
+
+- [x] HTTP Request, HTTP Response 클래스로 나눠서 구현했다.
+- [x] Controller 인터페이스와 RequestMapping 클래스를 활용하여 if절을 제거했다.
 
 ### 정적 파일 서빙
