+++ tomcat/src/main/java/nextstep/jwp/controller/LoginController.java
@@ -1,16 +1,17 @@
-package nextstep.jwp;
+package nextstep.jwp.controller;
 
 import java.util.HashMap;
 import java.util.Map;
-import org.apache.coyote.http11.AbstractController;
-import org.apache.coyote.http11.Session;
-import org.apache.coyote.http11.SessionStorage;
+import nextstep.jwp.service.LoginService;
+import org.apache.catalina.servlets.AbstractController;
+import org.apache.coyote.http11.Resource;
+import org.apache.coyote.http11.ResourceLocator;
+import org.apache.coyote.http11.general.HttpHeaders;
 import org.apache.coyote.http11.request.HttpRequest;
-import org.apache.coyote.http11.request.Resource;
-import org.apache.coyote.http11.request.ResourceLocator;
-import org.apache.coyote.http11.request.spec.HttpHeaders;
 import org.apache.coyote.http11.response.HttpResponse;
 import org.apache.coyote.http11.response.spec.HttpStatus;
+import org.apache.coyote.http11.session.Session;
+import org.apache.coyote.http11.session.SessionStorage;
 
 public class LoginController extends AbstractController {
