+++ tomcat/src/main/java/nextstep/jwp/controller/RegisterController.java
@@ -1,11 +1,12 @@
-package nextstep.jwp;
+package nextstep.jwp.controller;
 
 import java.util.HashMap;
 import java.util.Map;
-import org.apache.coyote.http11.AbstractController;
+import nextstep.jwp.service.LoginService;
+import org.apache.catalina.servlets.AbstractController;
+import org.apache.coyote.http11.ResourceLocator;
+import org.apache.coyote.http11.general.HttpHeaders;
 import org.apache.coyote.http11.request.HttpRequest;
-import org.apache.coyote.http11.request.ResourceLocator;
-import org.apache.coyote.http11.request.spec.HttpHeaders;
 import org.apache.coyote.http11.response.HttpResponse;
 import org.apache.coyote.http11.response.spec.HttpStatus;
