+++ tomcat/src/main/java/nextstep/jwp/controller/StaticResourceController.java
@@ -1,8 +1,8 @@
-package nextstep.jwp;
+package nextstep.jwp.controller;
 
-import org.apache.coyote.http11.AbstractController;
+import org.apache.catalina.servlets.AbstractController;
+import org.apache.coyote.http11.ResourceLocator;
 import org.apache.coyote.http11.request.HttpRequest;
-import org.apache.coyote.http11.request.ResourceLocator;
 import org.apache.coyote.http11.response.HttpResponse;
 
 public class StaticResourceController extends AbstractController {
