+++ tomcat/src/main/java/org/apache/catalina/connector/Connector.java
@@ -5,14 +5,14 @@ import java.io.UncheckedIOException;
 import java.net.ServerSocket;
 import java.net.Socket;
 import java.util.List;
-import nextstep.jwp.LoginController;
-import nextstep.jwp.RegisterController;
-import nextstep.jwp.RootController;
-import nextstep.jwp.StaticResourceController;
-import org.apache.coyote.ControllerMappings;
+import nextstep.jwp.controller.LoginController;
+import nextstep.jwp.controller.RegisterController;
+import nextstep.jwp.controller.RootController;
+import nextstep.jwp.controller.StaticResourceController;
+import org.apache.catalina.servlets.ControllerMappings;
 import org.apache.coyote.WebConfig;
 import org.apache.coyote.http11.Http11Processor;
-import org.apache.coyote.http11.request.ResourceLocator;
+import org.apache.coyote.http11.ResourceLocator;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
