+++ tomcat/src/main/java/org/apache/coyote/http11/Http11Processor.java
@@ -10,14 +10,12 @@ import java.nio.charset.StandardCharsets;
 import java.util.ArrayList;
 import java.util.List;
 import nextstep.jwp.exception.UncheckedServletException;
-import org.apache.coyote.Controller;
-import org.apache.coyote.ControllerMappings;
+import org.apache.catalina.servlets.Controller;
+import org.apache.catalina.servlets.ControllerMappings;
 import org.apache.coyote.Processor;
 import org.apache.coyote.WebConfig;
+import org.apache.coyote.http11.general.HttpHeaders;
 import org.apache.coyote.http11.request.HttpRequest;
-import org.apache.coyote.http11.request.Resource;
-import org.apache.coyote.http11.request.ResourceLocator;
-import org.apache.coyote.http11.request.spec.HttpHeaders;
 import org.apache.coyote.http11.request.spec.StartLine;
 import org.apache.coyote.http11.response.HttpResponse;
 import org.apache.coyote.http11.response.spec.HttpStatus;
