+++ tomcat/src/main/java/org/apache/coyote/http11/session/Cookie.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.http11.session;
 
 public class Cookie {
